requirements.txt
requests
beautifulsoup4
lxml
pandas
fake-useragent
python-dotenv
loguru
httpx
fastapi
uvicorn